refactor(openapi): promote response-gen data types to foundation — Phase 3a
Prepares the ground for moving \`openapi/route.rs\`, \`openapi/response/\`,
and \`openapi/response_trace.rs\` to mockforge-openapi by pulling the
pure-data types they depend on into mockforge-foundation so leaf crates
can import them without creating a cycle back into mockforge-core.

Promoted modules (core -> foundation, core keeps re-export shims):

- \`ai_response\` — \`AiResponseConfig\`, \`AiResponseMode\`,
  \`RequestContext\`, \`expand_prompt_template\`. No internal deps. Tests
  cross-check against \`mockforge-template-expansion\` via a dev-only dep.
- \`schema_diff\` — \`ValidationError\`, \`SchemaInfo\`, \`validation_diff\`,
  \`to_422_json\`, \`to_enhanced_422_json\`, etc. Pure serde types.
- \`response_generation_trace\` — \`ResponseGenerationTrace\` +
  \`BlendingDecision\`/\`FieldBlendingDecision\`/\`PersonaGraphNodeUsage\`/
  \`RuleExecution\`/\`TemplateExpansion\`. Previously at
  \`crate::reality_continuum::response_trace\`; core retains that legacy
  path via a re-export submodule.
- \`response_selection\` — \`ResponseSelectionMode\` + \`ResponseSelector\`.
  Moved out of mockforge-openapi into foundation since it's a generic
  response-variant selection primitive also used by non-OpenAPI response
  trace code. mockforge-openapi re-exports foundation's version so the
  \`mockforge_openapi::response_selection::\*\` paths continue to resolve.

What stays in core (moves to mockforge-openapi in Phase 3b):

- \`openapi/route.rs\`, \`openapi/response/\`, \`openapi/response_trace.rs\` —
  these still reference core-only types (\`MockAI\` concrete,
  \`template_expansion\`, \`templating\`). Phase 3b will introduce
  \`TemplateRenderer\`/\`Expander\` hook traits so the response generator
  can be lifted out of core.

No breaking changes. All existing import paths (e.g.
\`mockforge_core::schema_diff::ValidationError\`,
\`mockforge_core::reality_continuum::response_trace::ResponseGenerationTrace\`,
\`mockforge_core::ai_response::AiResponseConfig\`) resolve via re-exports.

Verification:
- cargo build --workspace --exclude mockforge-desktop clean
- cargo test -p mockforge-foundation --lib — 84 passed (up from 42; picked up the 42 tests that moved in)
- cargo test -p mockforge-openapi --lib — 66 passed
- cargo test -p mockforge-core --lib — 1710 passed (down from 1746; 36 tests moved out with the modules)
- cargo clippy --all-targets --all-features --workspace --exclude mockforge-desktop -- -D warnings clean
- cargo fmt --all --check clean

Plan: ~/.claude/plans/openapi-spec-extraction.md (Phase 3a)
