+++ crates/mockforge-foundation/src/response_selection.rs
