+++ crates/mockforge-foundation/src/schema_diff.rs
